Update to language-python 0.5.4.
This commit is a bit of a mess. language-python made a lot of changes to how
the AST is structured, in particular how 'x.y' is represented, which needed
lots of changes in how expression/variable binding works. Because of the
changes to binding, the basic lambda example now works!

It also added Functor definitions for the AST which means that most of
Normalize.hs is unneeded.